refactor: :construction: Refactor types in tasks and SH handler to pave way for a separation into its own crate
