+++ node/src/services/mod.rs
@@ -1,3 +1,4 @@
 pub mod builder;
 pub mod forest_storage;
 pub mod handler;
+pub mod types;
